[LIVEOAK-36] Support traversing URLs with matrix parameters in the case of ;config, specifically.
This involved reworking the traversing-responder classes, and
actually creating a more-generic TraversalPlan with a series
of Steps, which may be replaced to accomodate upsert semantics.

[LIVEOAK-31] Provide a ConfigurableResource interface which may
be mixed into any other Resource to allow access to a method

	Resource configuration()

which is slurped off when a ;config step is found during
traversal.  The configuration Resource's parent should be set
to that of its parent, and its ID should be exactly ';config'.

This should probably be refactored and cleaned up a tad.