[LIVEOAK-2] Wire up interceptors.
Many interceptors may now be installed immediately after decoding
the resource request (or before encoding the response).  Every
interceptor is bi-directional, but can choose to ignore either
half of the interaction should it like.

An interceptor may replace/alter the request and/or the response,
and may decide to send a response on its own, short-circuiting the
remainder of the container processing.